backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# HTTP客户端
httpx==0.25.2

# Redis
redis==5.0.1

# 向量数据库
chromadb==0.4.18

# Embedding模型
sentence-transformers==2.2.2

# 文档解析
pypdf2==3.0.1
python-docx==1.1.0

# 工具
python-dotenv==1.0.0
